cfm: Update with review comments

+++ offload.h
@@ -21,12 +21,12 @@ int cfm_offload_mep_config(uint32_t br_ifindex, uint32_t instance, struct mac_ad
 			   uint32_t mepid, uint16_t vid);
 int cfm_offload_mep_cnt_clear(uint32_t br_ifindex, uint32_t instance);
 int cfm_offload_cc_config(uint32_t br_ifindex, uint32_t instance, uint32_t enable,
-			  uint32_t interval, uint32_t priority, struct maid_data *maid);
+			  uint32_t interval, uint8_t priority, struct maid_data *maid);
 int cfm_offload_cc_rdi(uint32_t br_ifindex, uint32_t instance, uint32_t rdi);
 int cfm_offload_cc_peer(uint32_t br_ifindex, uint32_t instance, uint32_t remove, uint32_t mepid);
-int cfm_offload_cc_ccm_tx(uint32_t br_ifindex, uint32_t instance, uint32_t priority, uint32_t dei,
+int cfm_offload_cc_ccm_tx(uint32_t br_ifindex, uint32_t instance, uint8_t priority, uint32_t dei,
 			  struct mac_addr *dmac, uint32_t sequence, uint32_t period, uint32_t iftlv,
-			  uint32_t iftlv_value, uint32_t porttlv, uint32_t porttlv_value);
+			  uint8_t iftlv_value, uint32_t porttlv, uint8_t porttlv_value);
 int cfm_offload_cc_cnt_clear(uint32_t br_ifindex, uint32_t instance);
 
 int cfm_offload_init(void);
